Fix: refactor the package of repository
