+++ src/main/java/com/project/Project/repository/review/ReviewCategoryRepository.java
@@ -1,4 +1,4 @@
-package com.project.Project.repository;
+package com.project.Project.repository.review;
 
 import com.project.Project.domain.enums.ReviewCategoryEnum;
 import com.project.Project.domain.review.ReviewCategory;
