+++ src/main/java/com/project/Project/repository/review/ReviewImageRepository.java
@@ -1,4 +1,4 @@
-package com.project.Project.repository;
+package com.project.Project.repository.review;
 
 import com.project.Project.domain.review.ReviewImage;
 import org.springframework.data.jpa.repository.JpaRepository;
