+++ src/main/java/com/project/Project/repository/review/ReviewRepository.java
@@ -1,10 +1,9 @@
-package com.project.Project.repository;
+package com.project.Project.repository.review;
 
 import com.project.Project.domain.review.Review;
 import com.project.Project.domain.room.Room;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
 
 import java.util.List;
