+++ src/main/java/com/project/Project/serializer/review/ReviewSerializer.java
@@ -6,8 +6,8 @@ import com.project.Project.domain.embedded.AnonymousStatus;
 import com.project.Project.domain.enums.*;
 import com.project.Project.domain.review.*;
 import com.project.Project.domain.room.Room;
-import com.project.Project.repository.ReviewCategoryRepository;
-import com.project.Project.repository.ReviewKeywordRepository;
+import com.project.Project.repository.review.ReviewCategoryRepository;
+import com.project.Project.repository.review.ReviewKeywordRepository;
 import com.project.Project.service.impl.FileProcessService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -15,7 +15,10 @@ import org.springframework.web.multipart.MultipartFile;
 
 import javax.annotation.PostConstruct;
 import java.math.BigDecimal;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Locale;
 import java.util.stream.Collectors;
 
 @Component
@@ -45,12 +48,12 @@ public class ReviewSerializer {
         ArrayList<ReviewCategory> allReviewCategory = (ArrayList) staticReviewCategoryRepository.findAll();
 
         List<ReviewToReviewCategory> reviewToReviewCategoryList = Arrays.stream(request.getClass().getDeclaredFields()).filter(field -> ReviewCategoryEnum.contains(field.getName()))
-                .map((field)-> {
+                .map((field) -> {
                     field.setAccessible(true);
                     try {
                         BigDecimal score = (BigDecimal) field.get(request);
                         ArrayList<ReviewCategory> clonedAllReviewCategory = (ArrayList) allReviewCategory.clone();
-                        ReviewCategory targetReviewCategory = clonedAllReviewCategory.stream().filter(reviewCategory -> reviewCategory.getType().equals(ReviewCategoryEnum.valueOf(field.getName().toUpperCase(Locale.ROOT)))).findFirst().orElseThrow(()-> new RuntimeException());
+                        ReviewCategory targetReviewCategory = clonedAllReviewCategory.stream().filter(reviewCategory -> reviewCategory.getType().equals(ReviewCategoryEnum.valueOf(field.getName().toUpperCase(Locale.ROOT)))).findFirst().orElseThrow(() -> new RuntimeException());
                         ReviewToReviewCategory temp = ReviewToReviewCategory
                                 .builder()
                                 .score(score)
@@ -75,7 +78,7 @@ public class ReviewSerializer {
                 .filter((advantageKeyword) -> advantageKeywordEnumList.contains(advantageKeyword.getKeywordType()))
                 .map((reviewKeyword) -> {
                     ReviewToReviewKeyword temp =
-                    ReviewToReviewKeyword.builder().build();
+                            ReviewToReviewKeyword.builder().build();
                     temp.setReviewKeyword(reviewKeyword);
                     return temp;
                 }).collect(Collectors.toList());
@@ -115,7 +118,7 @@ public class ReviewSerializer {
 
         //ReviewKeywordList에 Review 할당
         selectedReviewAdvantageKeywordList.forEach((selectedReviewKeyword) -> selectedReviewKeyword.setReview(review));
-        selectedReviewDisadvantageKeywordList.forEach((selectedReviewDisadvantageKeyword)-> selectedReviewDisadvantageKeyword.setReview(review));
+        selectedReviewDisadvantageKeywordList.forEach((selectedReviewDisadvantageKeyword) -> selectedReviewDisadvantageKeyword.setReview(review));
 
         // ReviewToReviewCategory 연관관계
         reviewToReviewCategoryList.forEach((reviewToReviewCategory -> reviewToReviewCategory.setReview(review)));
