+++ src/main/java/com/project/Project/service/impl/ReviewImageServiceImpl.java
@@ -3,7 +3,7 @@ package com.project.Project.service.impl;
 import com.project.Project.domain.enums.FileFolder;
 import com.project.Project.domain.review.Review;
 import com.project.Project.domain.review.ReviewImage;
-import com.project.Project.repository.ReviewImageRepository;
+import com.project.Project.repository.review.ReviewImageRepository;
 import com.project.Project.service.ReviewImageService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -22,10 +22,10 @@ public class ReviewImageServiceImpl implements ReviewImageService {
 
     @Transactional
     public void saveImageList(List<MultipartFile> imageFileList, Review review) {
-        for(MultipartFile multipartFile : imageFileList) {
+        for (MultipartFile multipartFile : imageFileList) {
             String url = fileProcessService.uploadImage(multipartFile, FileFolder.REVIEW_IMAGES);
             ReviewImage reviewImage = ReviewImage.builder().url(url).build();
-                reviewImage.setReview(review);
+            reviewImage.setReview(review);
             reviewImageRepository.save(reviewImage);
         }
     }
