+++ src/test/java/com/project/Project/controller/validation/ReviewRequestValidationTest.java
@@ -5,7 +5,7 @@
 //import com.project.Project.config.auth.SecurityConfig;
 //import com.project.Project.controller.review.controller.ReviewRestController;
 //import com.project.Project.repository.building.BuildingRepository;
-//import com.project.Project.repository.ReviewRepository;
+//import com.project.Project.repository.review.ReviewRepository;
 //import com.project.Project.repository.RoomRepository;
 //import com.project.Project.service.BuildingService;
 //import com.project.Project.service.ReviewService;
